pyproject.toml
[project]
name = "entari-plugin-user"
version = "0.1.0"
description = "Entari 用户插件"
authors = [
    { name = "KomoriDev ", email = "mute231010@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = []

requires-python = ">=3.10"

dependencies = [
    "arclet-entari>=0.16.5",
    "entari-plugin-database>=0.2.1",
    "expiringdictx>=1.1.0",
]

[project.urls]
homepage = "https://github.com/netsora/entari-plugin-user"
repository = "https://github.com/netsora/entari-plugin-user"

[dependency-groups]
dev = [
    "arclet-entari[full]>=0.16.5",
    "entari-plugin-server>=0.5.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.0",
    "satori-python-adapter-onebot11>=0.2.5",
    "satori-python-adapter-satori>=0.2.3",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"

typeCheckingMode = "standard"
reportPrivateImportUsage = false
reportShadowedImports = false
disableBytesTypePromotions = true

[build-system]
requires = ["uv_build >=0.8.3, <0.9.0"]
build-backend = "uv_build"
